requirements.txt
nose==1.3.7
coverage==4.5.3
mock==3.0.4
fakeredis==1.0.3
google-cloud-pubsub==0.38.0
